Move app models into heroku_connect.models, move errors to heroku_connect.errors
