+++ heroku_connect/db/models/__init__.py
@@ -1,5 +1,3 @@
 from .base import *  # NoQA
-from .errors import *  # NoQA
 from .fields import *  # NoQA
 from .related import *  # NoQA
-from .triggerlog import *  # NoQA
